moved login stuff to userController
